requirements.txt
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
numpy==1.24.4
matplotlib==3.9.0
scikit-image==0.24.0
scipy==1.13.1
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2
albumentations==1.4.15
rasterio==1.3.10
timm==0.4.12
PyYAML==6.0.2
loguru==0.7.2
scikit-learn==1.5.2